Result view
When a player dies (suicide, killed, ran into wall), the player is added to a "death pool", and a result view is shown. The result view contains:

* A restart button (respawns the snake and renders the controls on the player's device)
* High score
* Latest score
* A form that submits the high score with a phone number

(Also made the snake flash when it spawns.)

+++ app/services/food-service.js
@@ -34,7 +34,12 @@ class FoodService {
 
             if (food.type === ServerConfig.FOOD.SWAP.TYPE && playerContainer.getNumberOfActivePlayers() > 1) {
                 const otherPlayer = playerContainer.getAnActivePlayer(playerWhoConsumedFood.id);
-                this.boardOccupancyService.removePlayerOccupancy(otherPlayer.id, otherPlayer.getSegments());
+
+                if (!playerContainer.isSpectating(otherPlayer.id)
+                    && !playerContainer.isDead(otherPlayer.id)
+                    && !playerContainer.isWaitingToRespawn(otherPlayer.id)) {
+                    this.boardOccupancyService.removePlayerOccupancy(otherPlayer.id, otherPlayer.getSegments());
+                }
                 this.boardOccupancyService.removePlayerOccupancy(playerWhoConsumedFood.id, playerWhoConsumedFood.getSegments());
                 const otherPlayerDirection = otherPlayer.direction;
                 const otherPlayerDirectionBeforeMove = otherPlayer.directionBeforeMove;
